Apply comment
