+++ libs/chat-shared/src/index.ts
@@ -2,4 +2,5 @@ export * from './models/chat.js';
 export * from './models/theme.js';
 export * from './models/auth.js';
 export * from './models/dial-model.js';
+export * from './types/attachment.js';
 export * from './utils/merge-class.js';
