+++ libs/chat-shared/src/models/chat.ts
@@ -1,3 +1,5 @@
+import { AttachmentType } from '../types/attachment.js';
+
 /** Metadata returned by the DIAL file/conversation listing API for a single resource node. */
 export interface ConversationMetadata {
   /** Display name of the resource. */
@@ -65,18 +67,6 @@ export interface StreamChunk {
     /** Zero-based index of this choice. */
     index: number;
   }>;
-}
-
-/** Discriminates the kind of content an attachment carries. */
-export enum AttachmentType {
-  /** Generic file attachment (non-image). */
-  File = 'file',
-  /** Raster or vector image. */
-  Image = 'image',
-  /** A saved prompt snippet. */
-  Prompt = 'prompt',
-  /** Text pasted directly by the user. */
-  Pasted = 'pasted',
 }
 
 /** Generic async-operation status, reusable for any request lifecycle. */
